requirements_agents.txt
# Existing
pyyaml
requests
pandas
plotly
playwright
playwright-stealth
beautifulsoup4

# Agent Optimization
fastapi==0.115.0
uvicorn[standard]==0.30.0
aiohttp==3.10.0
tenacity==9.0.0
sqlalchemy==2.0.35
asyncpg==0.29.0
redis==5.1.0
pydantic==2.9.0
jinja2==3.1.4
click==8.1.7
pytest-asyncio==0.24.0
httpx==0.27.0
mcp==1.0.0
scikit-learn==1.4.0
scipy==1.12.0
